refactor: simplify ordersTable module (#5908)
* chore: fix hardcoded ids

* refactor: simplify ordersTable module

* chore: fix lint

* chore: fix useOrderActions

* chore: fix useAllTransactions memo
