+++ apps/cowswap-frontend/src/modules/orders/hooks/usePendingOrdersPrices.ts
@@ -1,0 +1,7 @@
+import { useAtomValue } from 'jotai'
+
+import { PendingOrdersPrices, pendingOrdersPricesAtom } from '../state/pendingOrdersPricesAtom'
+
+export function usePendingOrdersPrices(): PendingOrdersPrices {
+  return useAtomValue(pendingOrdersPricesAtom)
+}
