+++ apps/cowswap-frontend/src/modules/orders/index.ts
@@ -1,8 +1,16 @@
 export * from './hooks/useTokensForOrdersList'
 export * from './hooks/useSWRProdOrders'
 export * from './utils/getTokensListFromOrders'
+
+export type { PendingOrderPrices, PendingOrdersPrices } from './state/pendingOrdersPricesAtom'
+export type { SpotPricesKeyParams } from './state/spotPricesAtom'
+
 export { TransactionContentWithLink } from './containers/TransactionContentWithLink'
 export { OrdersNotificationsUpdater } from './updaters/OrdersNotificationsUpdater'
+
+export { useGetSpotPrice } from './state/spotPricesAtom'
+export { usePendingOrdersPrices } from './hooks/usePendingOrdersPrices'
+
 export { emitPostedOrderEvent } from './utils/emitPostedOrderEvent'
 export { emitFulfilledOrderEvent } from './utils/emitFulfilledOrderEvent'
 export { emitCancelledOrderEvent } from './utils/emitCancelledOrderEvent'
