+++ apps/cowswap-frontend/src/modules/orders/state/spotPricesAtom.ts
@@ -58,9 +58,7 @@ function getSpotPrice(params: SpotPricesKeyParams, spotPrices: SpotPrices): Pric
   return marketInverted ? spotPrice.invert() : spotPrice
 }
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function useGetSpotPrice() {
+export function useGetSpotPrice(): (params: SpotPricesKeyParams) => Price<Currency, Currency> | null {
   const spotPrices = useAtomValue(spotPricesAtom)
 
   return useCallback((params: SpotPricesKeyParams) => getSpotPrice(params, spotPrices), [spotPrices])
