+++ apps/cowswap-frontend/src/modules/ordersTable/containers/OrdersReceiptModal/index.tsx
@@ -1,10 +1,12 @@
+import { ReactNode } from 'react'
+
 import { useENS } from '@cowprotocol/ens'
 import { useWalletInfo } from '@cowprotocol/wallet'
 import { CurrencyAmount } from '@uniswap/sdk-core'
 
 import JSBI from 'jsbi'
 
-import { PendingOrdersPrices } from 'modules/orders/state/pendingOrdersPricesAtom'
+import { PendingOrdersPrices } from 'modules/orders'
 import { useTwapOrderByChildId, useTwapOrderById } from 'modules/twap'
 
 import { calculatePrice } from 'utils/orderUtils/calculatePrice'
@@ -17,9 +19,7 @@ type OrdersReceiptModalProps = {
   pendingOrdersPrices: PendingOrdersPrices
 }
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function OrdersReceiptModal(props: OrdersReceiptModalProps) {
+export function OrdersReceiptModal(props: OrdersReceiptModalProps): ReactNode {
   // TODO: can we get selected order from URL by id?
   const order = useSelectedOrder()
   const { chainId } = useWalletInfo()
