+++ apps/cowswap-frontend/src/modules/ordersTable/containers/OrdersTableWidget/hooks/useOrdersTableList.ts
@@ -1,23 +1,15 @@
 import { useMemo } from 'react'
 
-import { Order, PENDING_STATES, OrderStatus } from 'legacy/state/orders/actions'
+import { Order, OrderStatus, PENDING_STATES } from 'legacy/state/orders/actions'
 import { useSetIsOrderUnfillable } from 'legacy/state/orders/hooks'
 
 import { getIsComposableCowOrder } from 'utils/orderUtils/getIsComposableCowOrder'
 import { getIsNotComposableCowOrder } from 'utils/orderUtils/getIsNotComposableCowOrder'
 
-import { TabOrderTypes } from '../../../types'
+import { OrdersTableList, OrderTableItem, TabOrderTypes } from '../../../types'
 import { getOrderParams } from '../../../utils/getOrderParams'
 import { groupOrdersTable } from '../../../utils/groupOrdersTable'
-import { getParsedOrderFromTableItem, isParsedOrder, OrderTableItem } from '../../../utils/orderTableGroupUtils'
-
-export interface OrdersTableList {
-  pending: OrderTableItem[]
-  history: OrderTableItem[]
-  unfillable: OrderTableItem[]
-  signing: OrderTableItem[]
-  all: OrderTableItem[]
-}
+import { getParsedOrderFromTableItem, isParsedOrder } from '../../../utils/orderTableGroupUtils'
 
 // TODO: Add proper return type annotation
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
@@ -31,7 +23,7 @@ const ordersSorter = (a: OrderTableItem, b: OrderTableItem) => {
 const ORDERS_LIMIT = 100
 
 // TODO: Break down this large function into smaller functions
-// eslint-disable-next-line max-lines-per-function
+
 export function useOrdersTableList(
   allOrders: Order[],
   orderType: TabOrderTypes,
@@ -104,7 +96,7 @@ export function useOrdersTableList(
 
           // Add to pending if in a pending state and not in signing state
           if (isPending && !isSigning) {
-            acc.pending.push(item)
+            acc.open.push(item)
           }
 
           // Add to history if not pending and not signing
@@ -114,7 +106,7 @@ export function useOrdersTableList(
 
           return acc
         },
-        { pending: [], history: [], unfillable: [], signing: [], all: [] },
+        { open: [], history: [], unfillable: [], signing: [], all: [] },
       ),
     [allSortedOrders, chainId, balancesAndAllowances, orderType, setIsOrderUnfillable],
   )
