+++ apps/cowswap-frontend/src/modules/ordersTable/containers/OrdersTableWidget/hooks/useValidatePageUrlParams.ts
@@ -4,14 +4,16 @@ import { useLocation } from 'react-router'
 
 import { useNavigate } from 'common/hooks/useNavigate'
 
-import { ORDERS_TABLE_PAGE_SIZE } from '../../../const/tabs'
+import { ORDERS_TABLE_PAGE_SIZE, OrderTabId } from '../../../const/tabs'
 import { buildOrdersTableUrl } from '../../../utils/buildOrdersTableUrl'
 import { parseOrdersTableUrl } from '../../../utils/parseOrdersTableUrl'
 
 // Reset page params if they are invalid
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function useValidatePageUrlParams(ordersLength: number, currentTabId: string, currentPageNumber: number) {
+export function useValidatePageUrlParams(
+  ordersLength: number,
+  currentTabId: OrderTabId,
+  currentPageNumber: number,
+): void {
   const location = useLocation()
   const navigate = useNavigate()
