+++ apps/cowswap-frontend/src/modules/ordersTable/hooks/useNavigateToOpenOrdersTable.ts
@@ -4,15 +4,13 @@ import { useNavigate } from 'common/hooks/useNavigate'
 
 import { useGetBuildOrdersTableUrl } from './useGetBuildOrdersTableUrl'
 
-import { OPEN_TAB } from '../const/tabs'
+import { OrderTabId } from '../const/tabs'
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function useNavigateToOpenOrdersTable() {
+export function useNavigateToOpenOrdersTable(): () => void {
   const navigate = useNavigate()
   const buildOrdersTableUrl = useGetBuildOrdersTableUrl()
 
   return useCallback(() => {
-    navigate(buildOrdersTableUrl({ tabId: OPEN_TAB.id, pageNumber: 1 }))
+    navigate(buildOrdersTableUrl({ tabId: OrderTabId.open, pageNumber: 1 }))
   }, [buildOrdersTableUrl, navigate])
 }
