+++ apps/cowswap-frontend/src/modules/ordersTable/hooks/useOrdersTableState.ts
@@ -1,0 +1,8 @@
+import { useAtomValue } from 'jotai'
+
+import { ordersTableStateAtom } from '../state/ordersTableStateAtom'
+import { OrdersTableState } from '../types'
+
+export function useOrdersTableState(): OrdersTableState | null {
+  return useAtomValue(ordersTableStateAtom)
+}
