+++ apps/cowswap-frontend/src/modules/ordersTable/pure/OrdersTableContainer/index.cosmos.tsx
@@ -1,23 +1,26 @@
+import { useSetAtom } from 'jotai/index'
+import { useEffect } from 'react'
+
+import { SupportedChainId } from '@cowprotocol/cow-sdk'
+
 import { BalancesAndAllowances } from 'modules/tokens'
 
 import { ParsedOrder } from 'utils/orderUtils/parseOrder'
 
 import { ordersMock } from './orders.mock'
-import { OrderActions } from './types'
 
-import { OrderTab } from '../../const/tabs'
-import { TabOrderTypes } from '../../types'
-
-import { OrdersTableContainer } from './index'
+import { OrderTab, OrderTabId } from '../../const/tabs'
+import { ordersTableStateAtom } from '../../state/ordersTableStateAtom'
+import { OrderActions, TabOrderTypes } from '../../types'
 
 const tabs: OrderTab[] = [
   {
-    id: 'open',
+    id: OrderTabId.open,
     title: 'Open orders',
     count: 5,
   },
   {
-    id: 'history',
+    id: OrderTabId.history,
     title: 'Orders history',
     count: 0,
     isActive: false,
@@ -56,24 +59,38 @@ const orderActions: OrderActions = {
   },
 }
 
-export default (
-  <OrdersTableContainer
-    pendingActivities={[]}
-    displayOrdersOnlyForSafeApp={false}
-    pendingOrdersPrices={{}}
-    chainId={1}
-    currentPageNumber={1}
-    orders={ordersMock}
-    tabs={tabs}
-    isSafeViaWc={false}
-    allowsOffchainSigning={true}
-    isWalletConnected={true}
-    selectedOrders={[]}
-    balancesAndAllowances={balancesAndAllowances}
-    getSpotPrice={() => null}
-    orderActions={orderActions}
-    orderType={TabOrderTypes.LIMIT}
-    injectedWidgetParams={{}}
-    isTwapTable={false}
-  />
-)
+function Wrapper(): null {
+  const setOrdersTableState = useSetAtom(ordersTableStateAtom)
+
+  useEffect(() => {
+    setOrdersTableState({
+      pendingActivities: [],
+      displayOrdersOnlyForSafeApp: false,
+      pendingOrdersPrices: {},
+      chainId: SupportedChainId.MAINNET,
+      currentPageNumber: 1,
+      orders: ordersMock,
+      filteredOrders: ordersMock,
+      tabs: tabs,
+      isSafeViaWc: false,
+      allowsOffchainSigning: true,
+      isWalletConnected: true,
+      selectedOrders: [],
+      balancesAndAllowances: balancesAndAllowances,
+      getSpotPrice: () => null,
+      orderActions: orderActions,
+      orderType: TabOrderTypes.LIMIT,
+      injectedWidgetParams: {},
+      isTwapTable: false,
+      currentTabId: OrderTabId.open,
+    })
+  }, [setOrdersTableState])
+
+  return null
+}
+
+const Fixtures = {
+  default: () => <Wrapper />,
+}
+
+export default Fixtures
