+++ apps/cowswap-frontend/src/modules/ordersTable/state/ordersTableStateAtom.ts
@@ -1,0 +1,5 @@
+import { atom } from 'jotai'
+
+import { OrdersTableState } from '../types'
+
+export const ordersTableStateAtom = atom<OrdersTableState | null>(null)
