+++ apps/cowswap-frontend/src/modules/ordersTable/utils/orderTableGroupUtils.ts
@@ -1,11 +1,6 @@
 import { ParsedOrder } from 'utils/orderUtils/parseOrder'
 
-export interface OrderTableGroup {
-  parent: ParsedOrder
-  children: ParsedOrder[]
-}
-
-export type OrderTableItem = OrderTableGroup | ParsedOrder
+import { OrderTableItem } from '../types'
 
 export const isParsedOrder = (item: OrderTableItem): item is ParsedOrder => !('children' in item)
