+++ apps/cowswap-frontend/src/modules/tokens/hooks/useBalancesAndAllowances.ts
@@ -1,0 +1,20 @@
+import { useMemo } from 'react'
+
+import { useTokensAllowances, useTokensBalances } from '@cowprotocol/balances-and-allowances'
+
+import type { BalancesAndAllowances } from '../types'
+
+export function useBalancesAndAllowances(): BalancesAndAllowances {
+  const balancesState = useTokensBalances()
+  const allowancesState = useTokensAllowances()
+
+  return useMemo(() => {
+    const { isLoading: balancesLoading, values: balances } = balancesState
+    const { isLoading: allowancesLoading, values: allowances } = allowancesState
+    return {
+      isLoading: balancesLoading || allowancesLoading,
+      balances,
+      allowances,
+    }
+  }, [balancesState, allowancesState])
+}
