+++ apps/cowswap-frontend/src/modules/tokens/index.ts
@@ -2,3 +2,4 @@ export * from './types'
 
 // Exported for all other cases, when we need the effective balance (i.e. )
 export * from './hooks/useEnoughBalance'
+export * from './hooks/useBalancesAndAllowances'
